+++ rust/blockstore/src/arrow/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod block;
 pub(crate) mod blockfile;
+#[cfg(test)]
 mod concurrency_test;
 pub mod config;
 pub(crate) mod flusher;
