+++ rust/blockstore/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod types;
 
 pub mod arrow;
+#[cfg(test)]
+mod blockfile_writer_test;
 pub mod config;
 pub mod key;
 pub mod memory;
