+++ rust/blockstore/src/types.rs
@@ -356,210 +356,3 @@ impl<
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::BTreeMap;
-
-    use chroma_storage::local::LocalStorage;
-    use chroma_storage::Storage;
-    use futures::executor::block_on;
-    use proptest::prelude::*;
-    use proptest::test_runner::Config;
-    use proptest_state_machine::prop_state_machine;
-    use proptest_state_machine::{ReferenceStateMachine, StateMachineTest};
-    use tempfile::TempDir;
-    use uuid::Uuid;
-
-    use crate::arrow::provider::ArrowBlockfileProvider;
-    use crate::BlockfileWriter;
-    use chroma_cache::new_cache_for_test;
-
-    #[derive(Clone, Debug)]
-    pub enum Transition {
-        Set(String, String, String),
-        Delete(String, String),
-        Commit,
-    }
-
-    #[derive(Debug, Clone)]
-    pub struct RefState {
-        /// This field is not used in the reference impl, but gives a block size to the real blockfile impl
-        generated_max_block_size_bytes: usize,
-        store: BTreeMap<(String, String), String>,
-        last_commit: Option<BTreeMap<(String, String), String>>,
-    }
-
-    pub struct BlockfileWriterStateMachine {}
-
-    impl ReferenceStateMachine for BlockfileWriterStateMachine {
-        type State = RefState;
-        type Transition = Transition;
-
-        fn init_state() -> proptest::prelude::BoxedStrategy<Self::State> {
-            (500..1_000usize)
-                .prop_map(|block_size_bytes| RefState {
-                    generated_max_block_size_bytes: block_size_bytes,
-                    store: BTreeMap::new(),
-                    last_commit: None,
-                })
-                .boxed()
-        }
-
-        fn transitions(state: &Self::State) -> proptest::prelude::BoxedStrategy<Self::Transition> {
-            let keys = state.store.keys().cloned().collect::<Vec<_>>();
-
-            let key_and_prefix_generator = (
-                "[0-9a-zA-Z]{1,10}",
-                "[0-9a-zA-Z]{1,10}",
-                "[0-9a-zA-Z]{1,100}",
-            );
-
-            if keys.is_empty() {
-                return key_and_prefix_generator
-                    .prop_map(|(prefix, key, value)| Transition::Set(prefix, key, value))
-                    .boxed();
-            }
-
-            prop_oneof![
-                4 => key_and_prefix_generator.prop_map(|(prefix, key, value)| {
-                    Transition::Set(prefix, key, value)
-                }),
-                2 => (0..keys.len()).prop_map(move |i| {
-                    Transition::Delete(keys[i].0.clone(), keys[i].1.clone())
-                }),
-                1 => Just(Transition::Commit)
-            ]
-            .boxed()
-        }
-
-        fn apply(mut state: Self::State, transition: &Self::Transition) -> Self::State {
-            match transition {
-                Transition::Set(prefix, key, value) => {
-                    state
-                        .store
-                        .insert((prefix.clone(), key.clone()), value.clone());
-                }
-                Transition::Delete(prefix, key) => {
-                    state.store.remove(&(prefix.clone(), key.clone())); // todo: need clones?
-                }
-                Transition::Commit => {
-                    state.last_commit = Some(state.store.clone());
-                }
-            }
-
-            state
-        }
-    }
-
-    struct BlockfileWriterWrapper {
-        tmp_dir: TempDir,
-        provider: ArrowBlockfileProvider,
-        last_blockfile_id: Option<Uuid>,
-        writer: BlockfileWriter,
-    }
-
-    impl StateMachineTest for BlockfileWriterWrapper {
-        type SystemUnderTest = Self;
-
-        type Reference = BlockfileWriterStateMachine;
-
-        fn init_test(
-            ref_state: &<Self::Reference as proptest_state_machine::ReferenceStateMachine>::State,
-        ) -> Self::SystemUnderTest {
-            let tmp_dir = tempfile::tempdir().unwrap();
-            let storage = Storage::Local(LocalStorage::new(tmp_dir.path().to_str().unwrap()));
-            let block_cache = new_cache_for_test();
-            let sparse_index_cache = new_cache_for_test();
-            let provider = ArrowBlockfileProvider::new(
-                storage,
-                ref_state.generated_max_block_size_bytes,
-                block_cache,
-                sparse_index_cache,
-            );
-            let writer = provider.create::<&str, String>().unwrap();
-
-            BlockfileWriterWrapper {
-                tmp_dir,
-                provider,
-                last_blockfile_id: None,
-                writer,
-            }
-        }
-
-        fn apply(
-            mut state: Self::SystemUnderTest,
-            ref_state: &<Self::Reference as proptest_state_machine::ReferenceStateMachine>::State,
-            transition: <Self::Reference as proptest_state_machine::ReferenceStateMachine>::Transition,
-        ) -> Self::SystemUnderTest {
-            match transition {
-                Transition::Set(prefix, key, value) => {
-                    block_on(state.writer.set(prefix.as_str(), key.as_str(), value)).unwrap();
-                }
-                Transition::Delete(prefix, key) => {
-                    block_on(
-                        state
-                            .writer
-                            .delete::<&str, String>(prefix.as_str(), key.as_str()),
-                    )
-                    .unwrap();
-                }
-                Transition::Commit => {
-                    let id = state.writer.id();
-                    let flusher = block_on(state.writer.commit::<&str, String>()).unwrap();
-                    block_on(flusher.flush::<&str, String>()).unwrap();
-
-                    state.last_blockfile_id = Some(id);
-                    state.writer = block_on(state.provider.fork::<&str, String>(&id)).unwrap();
-                }
-            }
-
-            state
-        }
-
-        fn check_invariants(
-            state: &Self::SystemUnderTest,
-            ref_state: &<Self::Reference as ReferenceStateMachine>::State,
-        ) {
-            if ref_state.last_commit.is_none() || state.last_blockfile_id.is_none() {
-                return;
-            }
-
-            let ref_last_commit = ref_state.last_commit.as_ref().unwrap();
-            let last_blockfile_id = state.last_blockfile_id.unwrap();
-
-            let reader = block_on(state.provider.open::<&str, &str>(&last_blockfile_id)).unwrap();
-
-            // Check count
-            assert_eq!(block_on(reader.count()).unwrap(), ref_last_commit.len());
-
-            // Check that entries are ordered and match expected
-            if let Some(min_key) = ref_last_commit.keys().next() {
-                let all_entries =
-                    block_on(reader.get_gte(min_key.0.as_str(), min_key.1.as_str())).unwrap();
-
-                for (blockfile_entry, expected_entry) in
-                    all_entries.iter().zip(ref_last_commit.iter())
-                {
-                    assert_eq!(blockfile_entry.0, expected_entry.0 .1); // key matches
-                    assert_eq!(blockfile_entry.1, expected_entry.1); // value matches
-                }
-            }
-        }
-    }
-
-    prop_state_machine! {
-        #![proptest_config(Config::default())]
-
-        #[test]
-        fn blockfile_writer_test(
-            sequential
-            // The number of transitions to be generated for each case.
-            1..100
-            // Macro's boilerplate to separate the following identifier.
-            =>
-            // The name of the type that implements `StateMachineTest`.
-            BlockfileWriterWrapper
-        );
-    }
-}
